+++ src/02-section/start-to-end/gallery.js
@@ -3,16 +3,18 @@ const Image = require('../../image/Image')
 //okay, <gallery> is a xml-tag, with newline-seperated data, somehow pivoted by '|'...
 //all deities help us. truly -> https://en.wikipedia.org/wiki/Help:Gallery_tag
 // - not to be confused with https://en.wikipedia.org/wiki/Template:Gallery...
-const parseGallery = function(section) {
+const parseGallery = function (section, doc) {
   let wiki = section.wiki
   wiki = wiki.replace(/<gallery([^>]*?)>([\s\S]+?)<\/gallery>/g, (_, attrs, inside) => {
     let images = inside.split(/\n/g)
-    images = images.filter(str => str && str.trim() !== '')
+    images = images.filter((str) => str && str.trim() !== '')
     //parse the line, which has an image and sometimes a caption
-    images = images.map(str => {
+    images = images.map((str) => {
       let arr = str.split(/\|/)
       let obj = {
-        file: arr[0].trim()
+        file: arr[0].trim(),
+        lang: doc.language,
+        domain: doc.domain,
       }
       let img = new Image(obj).json()
       let caption = arr.slice(1).join('|')
@@ -26,7 +28,7 @@ const parseGallery = function(section) {
       section.templates.push({
         template: 'gallery',
         images: images,
-        pos: section.title
+        pos: section.title,
       })
     }
     return ''